Refactor dev mode to use React context and provider wrapper
