+++ app/layout.tsx
@@ -1,6 +1,7 @@
 import { Analytics } from "@vercel/analytics/next";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import type { Metadata } from "next";
+import DevModeProviderWrapper from "@/components/DevModeProviderWrapper";
 import "./globals.css";
 
 export const metadata: Metadata = {
@@ -16,7 +17,9 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={`antialiased`}>
-        {children}
+        <DevModeProviderWrapper>
+          {children}
+        </DevModeProviderWrapper>
         <Analytics />
         <SpeedInsights />
       </body>
